@@ -1,6 +1,6 @@
 module cpd
 
-go 1.24.6
+go 1.25.0
 
 require github.com/elastic/go-elasticsearch/v8 v8.19.0
 
